+++ src/services/manufactures.js
@@ -1,5 +1,6 @@
 const manufactureRepository = require("../repositories/manufactures");
 const { imageUpload } = require("../utils/image-kit");
+const { NotFoundError } = require ("../utils/request");
 
 exports.getManufactures = async (req) => {
   return Object.keys(req.query).length
@@ -21,6 +22,11 @@ exports.createManufacture = async (data, file) => {
 };
 
 exports.updateManufacture = async (id, data, file) => {
+  const existingManufacture = await manufactureRepository.getManufactureById(id);
+  if (!existingManufacture) {
+    throw new NotFoundError("Manufacture not found!");
+  }
+
   if (file?.logo) {
     data.logo = await imageUpload(file.logo);
   }
@@ -28,5 +34,10 @@ exports.updateManufacture = async (id, data, file) => {
 };
 
 exports.deleteManufactureById = async (id) => {
+  const existingManufacture = await manufactureRepository.getManufactureById(id);
+  if (!existingManufacture) {
+    throw new NotFoundError("Manufacture not found!");
+  }
+  
   return await manufactureRepository.deleteManufactureById(id);
 };
